refactor HasOneThroughAssociation
